+++ oss-android-sdk/src/main/java/com/alibaba/sdk/android/oss/internal/ResponseParsers.java
@@ -39,7 +39,6 @@ import java.util.ArrayList;
 import java.util.Iterator;
 import java.util.List;
 import java.util.Map;
-import java.util.zip.CheckedInputStream;
 
 /**
  * Created by zhouzhuo on 11/23/15.
